pull jordan

+++ Patissier-iOS-Objective-C/Profile/Product.h
@@ -15,7 +15,7 @@
 
 @property NSString *name;
 
-@property NSInteger price;
+@property double price;
 
 - (NSURL *) imageURL;
